feat: 단체 탈퇴 service 메서드 추가
